add tests for the Windows habitat plan
Adds a test PowersShell script to habitat/tests/ to run some simple
tests on executable version output and then run the functional specs
suite like the omnibus_test script does.

scripts/bk_tests/verify_win_plan.ps1 will perform a throwaway build of
the plan under a "ci" origin and then run the test script upon the built
package.

The verify pipeline was updated to run the verify_win_plan script. This
will run for every new PR and every new commit to a PR. (Not sure that
this is the right timing/trigger for testing the plan.)

Signed-off-by: Robb Kidd <robb@thekidds.org>